refactor: implement Go search engine and C++ DocumentManager

+++ src/ScintillaWrapper.cpp
@@ -1,18 +1,10 @@
 /**
  * ScintillaWrapper.cpp
- *
- * Implementation of the type-safe C++ wrapper for Scintilla.
- * This class abstracts away `scintilla_send_message` logic.
  */
 
 #include "ScintillaWrapper.h"
 
-// In a real build, we would include the actual Scintilla header here:
-// #include "scintilla.h"
-// #include "SciLexer.h"
-
-// For testing compilation without full GTK/Scintilla headers present in the environment,
-// we stub the necessary constants and functions.
+// Note: In a full integration, we would include "scintilla.h" directly.
 #ifndef SCI_GETTEXT
 #define SCI_GETTEXT 2182
 #define SCI_SETTEXT 2181
@@ -24,18 +16,16 @@
 #define SCI_GETREADONLY 2140
 #endif
 
-// Stub the C function we are wrapping.
-extern "C" long scintilla_send_message(ScintillaObject* sci, unsigned int iMessage, unsigned long wParam, long lParam);
+// Updated signature to match the actual Scintilla API requirements for 64-bit compatibility
+extern "C" intptr_t scintilla_send_message(ScintillaObject* sci, unsigned int iMessage, uintptr_t wParam, intptr_t lParam);
 
 namespace geany {
 
 ScintillaWrapper::ScintillaWrapper(ScintillaObject* sci_obj) : m_sci(sci_obj) {}
 
-ScintillaWrapper::~ScintillaWrapper() {
-    // We do not destroy the widget here; GTK manages the widget lifecycle.
-}
+ScintillaWrapper::~ScintillaWrapper() {}
 
-long ScintillaWrapper::SendEditor(unsigned int message, unsigned long wParam, long lParam) const {
+intptr_t ScintillaWrapper::SendEditor(unsigned int message, uintptr_t wParam, intptr_t lParam) const {
     if (!m_sci) return 0;
     return scintilla_send_message(m_sci, message, wParam, lParam);
 }
@@ -45,20 +35,20 @@ std::string ScintillaWrapper::GetText() const {
     if (length == 0) return "";
 
     std::vector<char> buffer(length + 1);
-    SendEditor(SCI_GETTEXT, length + 1, reinterpret_cast<long>(buffer.data()));
+    SendEditor(SCI_GETTEXT, length + 1, reinterpret_cast<intptr_t>(buffer.data()));
     return std::string(buffer.data());
 }
 
 void ScintillaWrapper::SetText(const std::string& text) {
-    SendEditor(SCI_SETTEXT, 0, reinterpret_cast<long>(text.c_str()));
+    SendEditor(SCI_SETTEXT, 0, reinterpret_cast<intptr_t>(text.c_str()));
 }
 
 int ScintillaWrapper::GetLineCount() const {
-    return SendEditor(SCI_GETLINECOUNT);
+    return static_cast<int>(SendEditor(SCI_GETLINECOUNT));
 }
 
 void ScintillaWrapper::AppendText(const std::string& text) {
-    SendEditor(SCI_APPENDTEXT, text.length(), reinterpret_cast<long>(text.c_str()));
+    SendEditor(SCI_APPENDTEXT, text.length(), reinterpret_cast<intptr_t>(text.c_str()));
 }
 
 void ScintillaWrapper::ClearAll() {
@@ -66,7 +56,7 @@ void ScintillaWrapper::ClearAll() {
 }
 
 int ScintillaWrapper::GetLength() const {
-    return SendEditor(SCI_GETLENGTH);
+    return static_cast<int>(SendEditor(SCI_GETLENGTH));
 }
 
 void ScintillaWrapper::SetReadOnly(bool readOnly) {
